Add SSL extension, tests are still missing
