--- src/main/kotlin/com/github/imflog/schema/registry/register/RegisterSchemasTask.kt
@@ -1,44 +1,0 @@
-package com.github.imflog.schema.registry.register
-
-import com.github.imflog.schema.registry.RegistryClientWrapper
-import org.gradle.api.DefaultTask
-import org.gradle.api.GradleScriptException
-import org.gradle.api.model.ObjectFactory
-import org.gradle.api.provider.ListProperty
-import org.gradle.api.provider.Property
-import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.TaskAction
-import javax.inject.Inject
-
-open class RegisterSchemasTask @Inject constructor(objects: ObjectFactory) : DefaultTask() {
-
-    companion object {
-        const val TASK_NAME = "registerSchemasTask"
-    }
-
-    init {
-        group = "registry"
-        description = "Register schemas in the registry"
-    }
-
-    @Input
-    val url: Property<String> = objects.property(String::class.java)
-
-    @Input
-    val basicAuth: Property<String> = objects.property(String::class.java)
-
-    @Input
-    val subjects: ListProperty<RegisterSubject> = objects.listProperty(RegisterSubject::class.java)
-
-    @TaskAction
-    fun registerSchemas() {
-        val errorCount = RegisterTaskAction(
-            RegistryClientWrapper.client(url.get(), basicAuth.get()),
-            project.rootDir,
-            subjects.get()
-        ).run()
-        if (errorCount > 0) {
-            throw GradleScriptException("$errorCount schemas not registered, see logs for details", Throwable())
-        }
-    }
-}
