+++ src/main/kotlin/com/github/imflog/schema/registry/security/SslExtension.kt
@@ -1,0 +1,11 @@
+package com.github.imflog.schema.registry.security
+
+import org.gradle.api.model.ObjectFactory
+import org.gradle.api.provider.MapProperty
+
+open class SslExtension(objects: ObjectFactory) {
+
+    val configs: MapProperty<String, Any> = objects.mapProperty(String::class.java, Any::class.java).apply {
+        convention(mapOf())
+    }
+}
